Fix build issues for cross-compilation and no-pcap builds
- packet_handler.h: Add missing stdint.h for uint32_t (no-pcap builds)
- prolink_thread.c: Guard pcap_breakloop with HAVE_PCAP
- prolink_thread.c: Mark 'pt' as unused to fix warning
- capture_afxdp.c: Fix signedness comparison warning
- Makefile: Add BPF_CFLAGS for cross-compilation support

+++ audio/capture_afxdp.c
@@ -282,7 +282,7 @@ void *capture_afxdp(void *arg) {
 
         /* Refill fill queue with saved addresses - must not leak frames */
         uint32_t idx_fq = 0;
-        while (xsk_ring_prod__reserve(&umem.fq, rcvd, &idx_fq) != (int)rcvd) {
+        while (xsk_ring_prod__reserve(&umem.fq, rcvd, &idx_fq) != rcvd) {
             /* Fill ring full, kernel hasn't consumed yet - brief spin */
             if (!g_running) break;
         }
